Fixes in CandidateSimMuonMatcher
CandidateSimMuonMatcher.h
- added method matchSimple() that matches simTracks and OMTF cands
without propagation - it works well for the single muon samples. The
problem with propagation is that it does not work for low pt muons (pt <
~3GeV). Added flag usePropagation to switch between matchSimple() and
match() with propagation.
- added phi folding for delta phi in ghostBust() (but it looks like it
does not change the results)
- in match() (version for the SimTrack), the MatchingResult is added
also for the track for which the propagation failed.
The changes should be still added to the match() version for the
tracking particle.